Time: 1 ms (43.88%) | Memory: 112.3 MB (99.17%) - LeetSync

+++ 134-gas-station/gas-station.cpp
@@ -8,7 +8,7 @@ public:
 
         for(int i=0;i<gas.size();i++)
         {   
-            long diff=gas[i]-cost[i];
+            int diff=gas[i]-cost[i];
             total+=diff;
             tank+=diff;
             if(tank<0)
